+++ demos/storybook/src/utils.ts
@@ -1,10 +1,20 @@
 import '@pxblue/angular-themes/theme.scss';
 import { CommonModule } from '@angular/common';
 import { Component, NgModule } from '@angular/core';
-import 'typeface-open-sans';
+import '@fontsource/open-sans/300.css';
+import '@fontsource/open-sans/300-italic.css';
+import '@fontsource/open-sans/400.css';
+import '@fontsource/open-sans/400-italic.css';
+import '@fontsource/open-sans/600.css';
+import '@fontsource/open-sans/600-italic.css';
+import '@fontsource/open-sans/700.css';
+import '@fontsource/open-sans/700-italic.css';
+import '@fontsource/open-sans/800.css';
+import '@fontsource/open-sans/800-italic.css';
 import { BrowserModule } from '@angular/platform-browser';
 import { COMPONENT_SECTION_NAME, README_STORY_NAME } from './constants';
 import * as Colors from '@pxblue/colors';
+import { getDirection } from '@pxblue/storybook-rtl-addon';
 
 let banner: HTMLElement;
 let prevUrl = '';
@@ -21,6 +31,7 @@ const setBannerStyle = (display: string): void => getBanner().setAttribute('styl
 
 export const showTopBanner = (): void => setBannerStyle('unset');
 export const hideTopBanner = (): void => setBannerStyle('none');
+export const invertRTL = (): string => (getDirection() === 'rtl' ? 'scaleX(-1)' : '');
 
 export const selectCanvasTab = (): void => {
     window.top.history.replaceState(null, '', window.top.location.href.replace(NOTES_PATH, STORY_PATH));