+++ demos/storybook/stories/dropdown-toolbar/with-custom-menu.stories.ts
@@ -5,13 +5,20 @@ import { action } from '@storybook/addon-actions';
 export const withCustomMenu = (): any => ({
     template: `
        <pxb-dropdown-toolbar [title]="title" [subtitle]="state.selected || subtitle">
-        <ng-container pxb-toolbar-menu>
-            <mat-nav-list>
-                <pxb-info-list-item [dense]=true (click)="click(); updateSubtitle('Atlanta', state)"><mat-icon [style.color]="colors.blue[500]" pxb-icon>business</mat-icon><span pxb-title>Atlanta</span></pxb-info-list-item>
-                <pxb-info-list-item [dense]=true (click)="click(); updateSubtitle('Pittsburgh', state)" [statusColor]="colors.red[500]"><mat-icon [style.color]="colors.red[500]" pxb-icon>house</mat-icon><span pxb-title>Pittsburgh</span></pxb-info-list-item>
-                <pxb-info-list-item [dense]=true (click)="click(); updateSubtitle('New York', state)"><mat-icon [style.color]="colors.blue[500]" pxb-icon>apartment</mat-icon><span pxb-title>New York</span></pxb-info-list-item>
-            </mat-nav-list>
-        </ng-container>
+          <mat-nav-list pxb-toolbar-menu [style.paddingTop.px]="0">
+            <pxb-info-list-item [dense]=true (click)="click(); updateSubtitle('Atlanta', state)">
+                <mat-icon pxb-icon [style.color]="colors.blue[500]" style="transform: unset">business</mat-icon>
+                <span pxb-title>Atlanta</span>
+            </pxb-info-list-item>
+            <pxb-info-list-item [dense]=true (click)="click(); updateSubtitle('Pittsburgh', state)" [statusColor]="colors.red[500]">
+                <mat-icon pxb-icon [style.color]="colors.red[500]" style="transform: unset">house</mat-icon>
+                <span pxb-title>Pittsburgh</span>
+            </pxb-info-list-item>
+            <pxb-info-list-item [dense]=true (click)="click(); updateSubtitle('New York', state)">
+                <mat-icon pxb-icon [style.color]="colors.blue[500]" style="transform: unset">apartment</mat-icon>
+                <span pxb-title>New York</span>
+            </pxb-info-list-item>
+          </mat-nav-list>
        </pxb-dropdown-toolbar>
     `,
     props: {