+++ demos/storybook/stories/hero/with-full-config.stories.ts
@@ -1,12 +1,14 @@
 import { boolean, color, text, number } from '@storybook/addon-knobs';
+import { invertRTL } from '../../src/utils';
 import * as Colors from '@pxblue/colors';
 
 export const withFullConfig = (): any => ({
     template: `
-        <pxb-hero [label]="label" [value]="value" [units]="units"
-            [iconBackgroundColor]="iconBg" [iconSize]="iconSize">
+        <pxb-hero [label]="label" [value]="value" [units]="units" [iconSize]="iconSize">
             <i pxb-primary [style.color]="iconColor" class="pxb-fan"></i>
-            <mat-icon *ngIf="showSecondary" secondary>trending_up</mat-icon>
+            <mat-icon pxb-secondary *ngIf="showSecondary" [style.transform]="invertRTL()">
+                trending_up
+            </mat-icon>
         </pxb-hero>
       `,
     props: {
@@ -16,6 +18,6 @@ export const withFullConfig = (): any => ({
         showSecondary: boolean('Show Secondary Icon', true),
         iconSize: number('iconSize', 36),
         iconColor: color('primary.style.color', Colors.white[50]),
-        iconBg: color('primary.style.backgroundColor', Colors.blue[500]),
+        invertRTL: invertRTL,
     },
 });