+++ demos/storybook/stories/info-list-item/with-left-content.stories.ts
@@ -1,4 +1,5 @@
 import * as Colors from '@pxblue/colors';
+import { getDirection } from '@pxblue/storybook-rtl-addon';
 
 export const withLeftContent = (): any => ({
     template: `
@@ -6,10 +7,18 @@ export const withLeftContent = (): any => ({
             <div pxb-title>Info List Item</div>
             <div pxb-subtitle>with a ChannelValue component to the left</div>
             <mat-icon [style.color]="colors.blue[500]" pxb-icon>battery_charging_full</mat-icon>
-            <pxb-channel-value [value]="15" units="A" pxb-left-content></pxb-channel-value>
+            <div pxb-left-content style="display: flex; flex-direction: column"
+                [style.marginRight.px]="direction() === 'rtl' ? 0 : 48"
+                [style.marginLeft.px]="direction() === 'rtl' ? 48 : 0">
+                <div class="mat-body-2">
+                    8:32 <strong>AM</strong>
+                </div>
+                <div class="mat-body-2" style="margin-top: -4px">11/21/21</div>
+            </div>
         </pxb-info-list-item>
       `,
     props: {
         colors: Colors,
+        direction: getDirection,
     },
 });