+++ demos/storybook/stories/info-list-item/with-right-content.stories.ts
@@ -2,8 +2,8 @@ import * as Colors from '@pxblue/colors';
 
 export const withRightContent = (): any => ({
     template: `
-        <pxb-info-list-item chevron="true">
-            <div pxb-title>Info LIst Item</div>
+        <pxb-info-list-item>
+            <div pxb-title>Info List Item</div>
             <div pxb-subtitle>with a ChannelValue component to the right</div>
             <mat-icon [style.color]="colors.blue[500]" pxb-icon>battery_charging_full</mat-icon>
             <pxb-channel-value [value]="15" units="A" pxb-right-content></pxb-channel-value>