+++ demos/storybook/stories/score-card/with-actions.stories.ts
@@ -3,13 +3,13 @@ import { action } from '@storybook/addon-actions';
 import { withCustomHeaderStyles } from './with-custom-header.stories';
 import * as Colors from '@pxblue/colors';
 
-export const demoActions = ['search', 'mail', 'notifications', 'list_alt', 'cloud', 'more_vert'];
+export const demoActions = ['search', 'mail', 'notifications', 'favorite', 'cloud', 'more_vert'];
 
 export const withActions = (): any => ({
     styles: [
         `${withCustomHeaderStyles}
         ::ng-deep .pxb-score-card-header {
-            background-color: ${Colors.red[500]}!important;
+            background-color: ${Colors.blue[500]}!important;
         }`,
     ],
     template: `