+++ demos/storybook/stories/score-card/with-score-badge.stories.ts
@@ -12,6 +12,10 @@ export const withScoreBadge = (): any => ({
         .sb-score-card-content mat-list-item {
             height: 36px!important;
         }
+        .sb-score-card-content mat-icon {
+            margin-right: 32px;
+            padding: 0;
+        }
         .rtl.sb-score-card-content mat-icon {
             margin-left: 32px;
             margin-right: 0;
@@ -19,12 +23,6 @@ export const withScoreBadge = (): any => ({
         }
         ::ng-deep .rtl .mat-list-text { 
             padding-right: 0!important;
-        }
-        .sb-score-card-content mat-icon {
-            margin-right: 32px;
-        }
-        .sb-score-card-content .mat-line {
-            font-weight: 600!important;
         }
     `,
     ],
@@ -39,19 +37,19 @@ export const withScoreBadge = (): any => ({
             <mat-icon pxb-action-items (click)="actionClick('more_vert')">more_vert</mat-icon>
             <mat-list pxb-body class="sb-score-card-content" [class.rtl]="direction() === 'rtl'">
                 <mat-list-item>
-                    <p mat-line>0 Alarms</p>
+                    <p mat-line style="font-weight: 400">0 Alarms</p>
                     <mat-icon mat-list-icon>notifications</mat-icon>
                 </mat-list-item>
                 <mat-list-item [style.color]="colors.blue[500]">
-                    <p mat-line>1 Event</p>
+                    <p mat-line style="font-weight: 600">1 Event</p>
                     <mat-icon mat-list-icon>list_alt</mat-icon>
                 </mat-list-item>
                 <mat-list-item>
-                    <p mat-line>Online</p>
+                    <p mat-line style="font-weight: 400">Online</p>
                     <mat-icon mat-list-icon>cloud</mat-icon>
                 </mat-list-item>
             </mat-list>
-            <pxb-hero pxb-badge [label]="'Grade'" [value]="'98'" [units]="'/100'" [iconSize]="72" [iconBackgroundColor]="colors.white[50]">
+            <pxb-hero pxb-badge [label]="'Grade'" [value]="'98'" [units]="'/100'" [iconSize]="72">
                 <i pxb-primary [style.color]="colors.green[500]" class="pxb-grade_a"></i>
             </pxb-hero>
             <pxb-info-list-item pxb-action-row chevron="true" hidePadding="true" dense="true"(click)="actionRowClick()">