+++ demos/storybook/stories/user-menu/with-menu-header.stories.ts
@@ -1,5 +1,6 @@
 import { text } from '@storybook/addon-knobs';
 import { items } from './with-basic-config.stories';
+import { invertRTL } from '../../src/utils';
 
 export const withMenuHeader = (): any => ({
     template: `
@@ -11,7 +12,7 @@ export const withMenuHeader = (): any => ({
             <mat-nav-list pxb-menu-body [style.paddingTop.px]="0">
                 <pxb-info-list-item *ngFor="let item of items" [dense]="true" 
                     (click)="open=false; item.onSelect();">
-                    <mat-icon pxb-icon>{{item.icon}}</mat-icon>
+                    <mat-icon pxb-icon [style.transform]="invertRTL()">{{item.icon}}</mat-icon>
                     <div pxb-title>{{item.title}}</div>
                 </pxb-info-list-item>
             </mat-nav-list>
@@ -22,5 +23,6 @@ export const withMenuHeader = (): any => ({
         items: items,
         menuTitle: text('menuTitle', 'Sample Title'),
         menuSubtitle: text('menuSubtitle', 'Sample subtitle'),
+        invertRTL: invertRTL,
     },
 });